\docType{data}
\name{accounts.template}
\alias{accounts.template}
\title{Template for producing the water accounts with \link{get.accounts}}
\format{Data frame containing the list of accounting terms to be produced by AWRA-R}
\description{
  Template for producing the water accounts with
  \link{get.accounts}
}
\keyword{datasets}

