\name{awrarcalib.initialise}
\alias{awrarcalib.initialise}
\title{Initialise states for the awrar model (see \link{awrar.run})}
\usage{
  awrarcalib.initialise(parameters)
}
\arguments{
  \item{parameters}{Vector of parameters}
}
\value{
  Initial value of the S reservoir filling level
}
\description{
  Initialise states for the awrar model (see
  \link{awrar.run})
}

