\name{checksubperiods}
\alias{checksubperiods}
\title{Check that a subperiod object is suitable for use}
\usage{
  checksubperiods(subper)
}
\arguments{
  \item{subper}{A subperiod object (e.g.
  \link{subperiods})}
}
\value{
  A list object containing two elements: \code{passed}
  indicates the test is passed, \code{message} gives some
  info
}
\description{
  Check that a subperiod object is suitable for use
}

