\name{get.default.config}
\alias{get.default.config}
\title{Get a default config parameter for \link{awrar.run}}
\usage{
  get.default.config()
}
\value{
  a config vector (see \link{awrar.run}) :
}
\description{
  The default config only activates routing and set the
  times step to daily
}
\examples{
cf <- get.default.config()
}

